test093-msgpack/py/requirements.txt
msgpack-python==0.4.8
numpy==1.13.3
